Refactor reordering
